Update customize_unsupported_guest_os.py
This is an example of customizing unsupported debian flavor. This is also an example of how to run any program inside guest OS. From 5.x onward the VIX API is merged with the vsphere mgmt. API. 

NB: The method StartProgramInGuest() returns only the process id. So if anything fails there is no way to figure out.